broke up rating trends util functions
