+++ src/components/Utilities/RatingTrendsValues.js
@@ -1,10 +1,18 @@
 import PropTypes from 'prop-types';
-import HelpIcon from '@mui/icons-material/Help';
 import {
-  Typography, Rating,
+  Typography,
 } from '@mui/material';
-import ToolTip from '@mui/material/Tooltip';
-import theme from '../../assets/styles/AppTheme';
+import {
+  starTitle,
+  compositePercentTitle,
+  measurePercentTitle,
+  submeasurePercentTitle,
+  percentDisplayValue,
+  starDisplayValue,
+  submeasurePercentDisplayValue,
+  percentageFooter,
+  starFooter
+} from './RatingTrendValuesUtil'
 import {
   activeMeasureProps, currentResultsProps, trendsProps, widgetPrefsProps, measureCheckerProps,
 } from './PropTypes';
@@ -13,96 +21,20 @@ export const ratingTrendsTip = 'Rating and Trends displays the current projected
 export const starsTip = 'Star rating subject to change depending on measures and other resources. For more information, please contact NCQA.';
 
 export function Title({ activeMeasure, preferences, currentResults }) {
-  // handles stars
   if (preferences?.type === 'star') {
-    return (
-      <Typography
-        variant="h6"
-        sx={{
-          padding: '1rem',
-          fontWeight: 700,
-          whiteSpace: 'nowrap',
-        }}
-      >
-        {`${preferences.measure.toUpperCase()} Star Rating`}
-        <ToolTip title={starsTip} sx={{ alignSelf: 'center' }}>
-          <HelpIcon color="secondary" className="rating-trends__help-icon" fontSize="small" />
-        </ToolTip>
-      </Typography>
-    )
+    return starTitle(preferences)
   }
-  // handles percentages
-  // composite percentages
   if (activeMeasure.measure === 'composite' && preferences?.type === 'percentage') {
-    return (
-      <Typography
-        variant="h6"
-        sx={{
-          padding: '1rem',
-          fontWeight: 700,
-          whiteSpace: 'nowrap',
-        }}
-      >
-        {`${preferences.measure.toUpperCase()} Score % Change`}
-        <ToolTip title={ratingTrendsTip} sx={{ alignSelf: 'center' }}>
-          <HelpIcon color="secondary" className="rating-trends__help-icon" fontSize="small" />
-        </ToolTip>
-      </Typography>
-    )
+    return compositePercentTitle(preferences)
   }
-  // measure percentages
   if (activeMeasure.measure === preferences.measure && preferences?.type === 'percentage') {
-    return (
-      <Typography
-        variant="h6"
-        sx={{
-          padding: '1rem',
-          fontWeight: 700,
-          whiteSpace: 'nowrap',
-        }}
-      >
-        {preferences.measure.toUpperCase()}
-        {' '}
-        Score % Change
-        <ToolTip title={ratingTrendsTip} sx={{ alignSelf: 'center' }}>
-          <HelpIcon color="secondary" className="rating-trends__help-icon" fontSize="small" />
-        </ToolTip>
-      </Typography>
-    )
+    return measurePercentTitle(preferences)
   }
   // sub-measure percentages submeasures
   if (activeMeasure.measure !== preferences.measure && preferences?.type === 'percentage') {
-    const subMeasures = currentResults.find(
-      (trend) => trend.measure === activeMeasure.measure,
-    ).subScores
-
-    let label = subMeasures?.find((sub) => preferences.measure === sub.measure).label
-    label = `${label.split('').slice(activeMeasure.measure.length + 4).join('')} Score % Change`
-
-    // submeasure titles can be super long so we're making the font size
-    // slightly smaller for massive ones
-    const charCount = label.split('').length > 45
-
-    return (
-      <Typography
-        variant="h6"
-        sx={{
-          padding: '.5rem',
-          fontWeight: 700,
-          width: '85%',
-          height: 'fit-content',
-          textAlign: 'center',
-          justifySelf: 'center',
-          fontSize: charCount ? '1rem' : '1.2rem',
-        }}
-      >
-        {label}
-        <ToolTip title={ratingTrendsTip} sx={{ alignSelf: 'center' }}>
-          <HelpIcon color="secondary" className="rating-trends__help-icon" fontSize="small" />
-        </ToolTip>
-      </Typography>
-    )
+    return submeasurePercentTitle(activeMeasure, preferences, currentResults)
   }
+
   return (
     <Typography
       variant="h6"
@@ -133,85 +65,32 @@ export function measureChecker(activeMeasure, preferences) {
   return measureCheck
 }
 
-export function WidgetValue({
+export function DisplayValue({
   activeMeasure, preferences, currentResults, trends, measureCheck,
 }) {
   const { percentCheck, starCheck, submeasureCheck } = measureCheck
 
   if (percentCheck) {
-    const percentValue = trends.find(
-      (trend) => trend.measure === preferences.measure.toLowerCase(),
-    ).percentChange
-    let percentColor = theme.palette?.text.disabled
-    if (percentValue > 0) {
-      percentColor = theme.palette?.success.main
-    } else if (percentValue < 0) {
-      percentColor = theme.palette?.error.main
-    }
-    return (
-      <Typography color={percentColor} variant="h4" sx={{ height: 'fit-content', padding: '0' }}>
-        {percentValue < 0 ? percentValue : `+ ${percentValue}`}
-        {' '}
-        %
-      </Typography>
-    )
+    return percentDisplayValue(trends, preferences)
   } else if (starCheck) {
-    const starValue = currentResults.find(
-      (trend) => trend.measure === preferences.measure.toLowerCase(),
-    ).starRating
-    return (
-      <Rating
-        name="read-only"
-        value={starValue}
-        precision={0.5}
-        sx={{ fontSize: 'xxx-large' }}
-        readOnly
-      />
-    )
+    return starDisplayValue(currentResults, preferences)
   } else if (submeasureCheck) {
-    const percentValue = trends.find(
-      (trend) => activeMeasure.measure === trend.measure,
-    ).subScoreTrends.find(
-      (trend) => trend.measure === preferences.measure,
-    ).percentChange
-    let percentColor = theme.palette?.text.disabled
-    if (percentValue > 0) {
-      percentColor = theme.palette?.success.main
-    } else if (percentValue < 0) {
-      percentColor = theme.palette?.error.main
-    }
-    return (
-      <Typography color={percentColor} variant="h4" sx={{ height: 'fit-content', padding: '0' }}>
-        {percentValue < 0 ? percentValue : `+${percentValue}`}
-        {' '}
-        %
-      </Typography>
-    )
+    return submeasurePercentDisplayValue(trends, activeMeasure, preferences)
   }
   return (
     <Typography>
-      Undefined component
+      Undefined value
     </Typography>
   )
 }
 
-export function Details({ preferences }) {
-  // props: measure, time period?
-  // either "over the past week" or measure title
+export function Footer({ preferences }) {
   if (preferences.type === 'percentage') {
-    return (
-      <Typography sx={{ height: 'fit-content' }}>
-        {preferences.measure.toUpperCase()}
-      </Typography>
-    )
+    return percentageFooter(preferences)
   } else if (preferences.type === 'star') {
-    return (
-      <Typography sx={{ height: '3rem', alignItems: 'center' }}>
-        (over the past week)
-      </Typography>
-    )
+    return starFooter()
   }
-  return undefined;
+  return '';
 }
 
 export const submeasureResults = (activeMeasure, trends) => {
@@ -252,14 +131,14 @@ Title.propTypes = {
   preferences: widgetPrefsProps,
   currentResults: PropTypes.arrayOf(PropTypes.shape({})),
 }
-WidgetValue.propTypes = {
+DisplayValue.propTypes = {
   activeMeasure: activeMeasureProps,
   currentResults: currentResultsProps,
   trends: trendsProps,
   preferences: widgetPrefsProps,
   measureCheck: measureCheckerProps,
 }
-Details.propTypes = {
+Footer.propTypes = {
   preferences: widgetPrefsProps,
 }
 
@@ -269,14 +148,14 @@ Title.defaultProps = {
   preferences: {},
   currentResults: {},
 }
-WidgetValue.defaultProps = {
+DisplayValue.defaultProps = {
   activeMeasure: {},
   currentResults: {},
   trends: {},
   preferences: {},
   measureCheck: {},
 }
-Details.defaultProps = {
+Footer.defaultProps = {
   preferences: {},
 }
 measureChecker.defaultProps = {
